refactor(cache): rename and restructure cache-related files
- Rename lv_cache_lru_rb.c and lv_cache_lru_rb.h to clazz directory
- Rename lv_image_cache.c and lv_image_cache.h to instance directory
- Rename lv_image_header_cache.c and lv_image_header_cache.h to instance directory
- Update include paths in affected files
- Replace lv_image_cache.h with lv_cache.h in files that use caching functionality
