+++ lvgl.h
@@ -36,7 +36,7 @@ extern "C" {
 #include "src/misc/lv_iter.h"
 #include "src/misc/lv_circle_buf.h"
 #include "src/misc/lv_tree.h"
-#include "src/misc/cache/lv_image_cache.h"
+#include "src/misc/cache/lv_cache.h"
 
 #include "src/tick/lv_tick.h"
