+++ src/misc/cache/clazz/lv_cache_clazz.h
@@ -1,0 +1,15 @@
+/**
+ * @file lv_cache_clazz.h
+ *
+ */
+
+#ifndef LV_CACHE_CLAZZ_H
+#define LV_CACHE_CLAZZ_H
+
+
+/*********************
+ *      INCLUDES
+ *********************/
+#include "lv_cache_lru_rb.h"
+
+#endif //LV_CACHE_CLAZZ_H
