+++ src/misc/cache/clazz/lv_cache_lru_rb.c
@@ -43,13 +43,15 @@
 /*********************
  *      INCLUDES
  *********************/
+
 #include "lv_cache_lru_rb.h"
-#include "../../stdlib/lv_sprintf.h"
-#include "../../stdlib/lv_string.h"
-#include "../lv_ll.h"
-#include "../lv_rb_private.h"
-#include "../lv_rb.h"
-#include "../lv_iter.h"
+#include "../lv_cache_entry.h"
+#include "../../../stdlib/lv_sprintf.h"
+#include "../../../stdlib/lv_string.h"
+#include "../../lv_ll.h"
+#include "../../lv_rb_private.h"
+#include "../../lv_rb.h"
+#include "../../lv_iter.h"
 
 /*********************
  *      DEFINES
