+++ src/misc/cache/clazz/lv_cache_lru_rb.h
@@ -13,8 +13,8 @@ extern "C" {
 /*********************
  *      INCLUDES
  *********************/
-#include "lv_cache_entry.h"
-#include "lv_cache_private.h"
+
+#include "../lv_cache_private.h"
 
 /*********************
  *      DEFINES
