+++ src/misc/cache/instance/lv_image_cache.c
@@ -7,13 +7,12 @@
  *      INCLUDES
  *********************/
 
-#include "../../draw/lv_image_decoder_private.h"
-#include "../lv_assert.h"
-#include "../../core/lv_global.h"
-#include "../../misc/lv_iter.h"
+#include "../../../draw/lv_image_decoder_private.h"
+#include "../../lv_assert.h"
+#include "../../../core/lv_global.h"
+#include "../../../misc/lv_iter.h"
 
 #include "lv_image_cache.h"
-#include "lv_image_header_cache.h"
 
 /*********************
  *      DEFINES
