+++ src/misc/cache/instance/lv_image_header_cache.h
@@ -14,8 +14,7 @@ extern "C" {
  *      INCLUDES
  *********************/
 
-#include "../../lv_conf_internal.h"
-#include "../lv_types.h"
+#include "../../lv_types.h"
 
 /*********************
  *      DEFINES
