+++ src/misc/lv_utils.c
@@ -9,7 +9,7 @@
 #include "lv_utils.h"
 #include "lv_fs.h"
 #include "lv_types.h"
-#include "cache/lv_image_cache.h"
+#include "cache/lv_cache.h"
 
 /*********************
  *      DEFINES
