+++ src/widgets/lottie/lv_lottie.c
@@ -18,7 +18,7 @@
 
 #include "../../misc/lv_timer.h"
 #include "../../core/lv_obj_class_private.h"
-#include "../../misc/cache/lv_image_cache.h"
+#include "../../misc/cache/lv_cache.h"
 
 /*********************
  *      DEFINES
